Rename `go mod` module and modify the import paths in each pkg
* Update module name in `go mod` and remove the line for referencing local module, `iamtokenvalidator`
* Delete `go.mod`and `go.sum` from the path, `/iamtokenvalidator`
* Modify the import paths in each pkg to align with the renamed module name, `github.com/m-cmp/mc-iam-manager`
