+++ actions/mcinfra.go
@@ -5,9 +5,9 @@ import (
 	"log"
 	"net/http"
 
-	"mc-iam-manager/handler"
-	"mc-iam-manager/handler/mcinframanager"
-	"mc-iam-manager/models"
+	"github.com/m-cmp/mc-iam-manager/handler"
+	"github.com/m-cmp/mc-iam-manager/handler/mcinframanager"
+	"github.com/m-cmp/mc-iam-manager/models"
 
 	"github.com/gobuffalo/buffalo"
 	"github.com/gobuffalo/nulls"
