+++ actions/workspace_user_role_mappings.go
@@ -4,8 +4,8 @@ import (
 	"log"
 	"net/http"
 
-	"mc-iam-manager/handler"
-	"mc-iam-manager/models"
+	"github.com/m-cmp/mc-iam-manager/handler"
+	"github.com/m-cmp/mc-iam-manager/models"
 
 	"github.com/gobuffalo/buffalo"
 	"github.com/gobuffalo/pop/v6"
