+++ handler/securitykeyprovider/alibaba/alibabaSecuritykeyprovider.go
@@ -8,7 +8,7 @@ import (
 	"net/http"
 	"os"
 
-	"mc-iam-manager/handler/securitykeyprovider"
+	"github.com/m-cmp/mc-iam-manager/handler/securitykeyprovider"
 
 	"github.com/gobuffalo/buffalo"
 )
