+++ handler/securitykeyprovider/aws/awsSecuritykeyprovider.go
@@ -7,7 +7,7 @@ import (
 	"net/http"
 	"os"
 
-	"mc-iam-manager/handler/securitykeyprovider"
+	"github.com/m-cmp/mc-iam-manager/handler/securitykeyprovider"
 
 	"github.com/gobuffalo/buffalo"
 )
