+++ handler/workspace_project_mappingsHandler.go
@@ -3,7 +3,7 @@ package handler
 import (
 	"strings"
 
-	"mc-iam-manager/models"
+	"github.com/m-cmp/mc-iam-manager/models"
 
 	"github.com/gobuffalo/pop/v6"
 	"github.com/gofrs/uuid"
