--- iamtokenvalidator/go.mod
@@ -1,20 +1,0 @@
-module github.com/m-cmp/mc-iam-manager/iamtokenvalidator
-
-go 1.21.6
-
-require (
-	github.com/golang-jwt/jwt/v4 v4.5.0
-	github.com/lestrrat-go/jwx v1.2.29
-)
-
-require (
-	github.com/decred/dcrd/dcrec/secp256k1/v4 v4.2.0 // indirect
-	github.com/goccy/go-json v0.10.2 // indirect
-	github.com/lestrrat-go/backoff/v2 v2.0.8 // indirect
-	github.com/lestrrat-go/blackmagic v1.0.2 // indirect
-	github.com/lestrrat-go/httpcc v1.0.1 // indirect
-	github.com/lestrrat-go/iter v1.0.2 // indirect
-	github.com/lestrrat-go/option v1.0.1 // indirect
-	github.com/pkg/errors v0.9.1 // indirect
-	golang.org/x/crypto v0.21.0 // indirect
-)
